Issue red #5083; * Improve QueueAllPageLocalizationsCommand;
git-svn-id: http://svn.vig/supra7/trunk@1730 f05b830f-2085-0410-a479-8f75d39d0b34

+++ src/lib/Supra/Controller/Pages/Search/PageLocalizationIndexerQueue.php
@@ -107,9 +107,13 @@ class PageLocalizationIndexerQueue extends IndexerQueue
 			);
 
 			$queueItem = $this->repository->findOneBy($criteria);
+			/* @var $queueItem PageLocalizationIndexerQueueItem */
 
 			if ( ! empty($queueItem)) {
-				$this->em->remove($queueItem);
+				
+				if ($queueItem->getRevisionId() != $pageLocalizationIndexerQueueItem->getRevisionId()) {
+					$this->em->remove($queueItem);
+				}
 			} else {
 				
 			}
@@ -118,4 +122,5 @@ class PageLocalizationIndexerQueue extends IndexerQueue
 		$pageLocalizationIndexerQueueItem->setSchemaName($this->schemaName);
 		parent::store($pageLocalizationIndexerQueueItem);
 	}
+
 }
